requirements.txt
Django==1.9.7
django-filter==0.13.0
djangorestframework==3.3.3
mysqlclient==1.3.7
requests==2.10.0
